requirements.txt
# ============================================================
# ml-fraud-credit-system — Python Dependencies
# Python 3.10+ required
# ============================================================

# ── Core packaging fix ─────────────────────────────────────
setuptools<70

# --- Core Data ---
pandas==2.2.2
numpy==1.26.4

# --- Database ---
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# --- ML & Feature Engineering ---
scikit-learn==1.5.0
xgboost==2.0.3
imbalanced-learn==0.12.3
joblib==1.4.2

# --- Experiment Tracking ---
mlflow==2.13.0

# --- API ---
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
python-multipart==0.0.9

# --- Monitoring ---
evidently==0.4.30
prometheus-fastapi-instrumentator==7.0.0
prometheus-client==0.20.0

# --- Utilities ---
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.27.0          # for async HTTP in tests

# --- Testing ---
pytest==8.2.2
pytest-asyncio==0.23.7

# --- Notebooks (optional, for EDA) ---
jupyter==1.0.0
matplotlib==3.9.0
seaborn==0.13.2


